refactor to leverage abstract tracker
