+++ lib/coverband.rb
@@ -16,6 +16,7 @@ require "coverband/adapters/stdout_store"
 require "coverband/adapters/null_store"
 require "coverband/utils/file_hasher"
 require "coverband/collectors/coverage"
+require "coverband/collectors/abstract_tracker"
 require "coverband/collectors/view_tracker"
 require "coverband/collectors/view_tracker_service"
 require "coverband/collectors/route_tracker"
