+++ test/coverband/collectors/view_tracker_test.rb
@@ -4,7 +4,7 @@ require File.expand_path("../../test_helper", File.dirname(__FILE__))
 
 class ViewTrackerTest < Minitest::Test
   def tracker_key
-    "render_tracker_2"
+    "ViewTracker_tracker"
   end
 
   def setup
@@ -18,7 +18,7 @@ class ViewTrackerTest < Minitest::Test
     assert_equal "dir", tracker.roots.first
     assert !tracker.store.nil?
     assert_equal [], tracker.target
-    assert_equal [], tracker.logged_views
+    assert_equal [], tracker.logged_keys
   end
 
   test "track partials" do
@@ -27,9 +27,9 @@ class ViewTrackerTest < Minitest::Test
     file_path = "#{File.expand_path(Coverband.configuration.root)}/file"
     store.raw_store.expects(:hset).with(tracker_key, file_path, anything)
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal [file_path], tracker.logged_views
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal [file_path], tracker.logged_keys
   end
 
   test "track partials that include the word vendor in the path" do
@@ -37,9 +37,9 @@ class ViewTrackerTest < Minitest::Test
     store = fake_store
     file_path = "#{File.expand_path(Coverband.configuration.root)}/vendor_relations/file"
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal [file_path], tracker.used_views.keys
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal [file_path], tracker.used_keys.keys
   end
 
   test "track partials that include the word _mailer in the path" do
@@ -47,9 +47,9 @@ class ViewTrackerTest < Minitest::Test
     store = fake_store
     file_path = "#{File.expand_path(Coverband.configuration.root)}/_mailer/file"
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal [file_path], tracker.used_views.keys
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal [file_path], tracker.used_keys.keys
   end
 
   test "ignore partials that include the folder vendor in the path" do
@@ -57,9 +57,9 @@ class ViewTrackerTest < Minitest::Test
     store = fake_store
     file_path = "#{File.expand_path(Coverband.configuration.root)}/vendor/file"
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal({}, tracker.used_views)
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal({}, tracker.used_keys)
   end
 
   test "track layouts" do
@@ -68,9 +68,9 @@ class ViewTrackerTest < Minitest::Test
     file_path = "#{File.expand_path(Coverband.configuration.root)}/layout"
     store.raw_store.expects(:hset).with(tracker_key, file_path, anything)
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", layout: file_path)
-    tracker.report_views_tracked
-    assert_equal [file_path], tracker.logged_views
+    tracker.track_key(layout: file_path)
+    tracker.save_report
+    assert_equal [file_path], tracker.logged_keys
   end
 
   test "report used partials" do
@@ -78,9 +78,9 @@ class ViewTrackerTest < Minitest::Test
     store = fake_store
     file_path = "#{File.expand_path(Coverband.configuration.root)}/file"
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal [file_path], tracker.used_views.keys
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal [file_path], tracker.used_keys.keys
   end
 
   test "report unused partials" do
@@ -89,30 +89,30 @@ class ViewTrackerTest < Minitest::Test
     file_path = "#{File.expand_path(Coverband.configuration.root)}/file"
     target = [file_path, "not_used"]
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir", target: target)
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.report_views_tracked
-    assert_equal ["not_used"], tracker.unused_views
+    tracker.track_key(identifier: file_path)
+    tracker.save_report
+    assert_equal ["not_used"], tracker.unused_keys
   end
 
   test "reset store" do
     Coverband::Collectors::ViewTracker.expects(:supported_version?).returns(true)
     store = fake_store
     store.raw_store.expects(:del).with(tracker_key)
-    store.raw_store.expects(:del).with("render_tracker_time")
+    store.raw_store.expects(:del).with("ViewTracker_tracker_time")
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: "file")
+    tracker.track_key(identifier: "file")
     tracker.reset_recordings
   end
 
-  test "clear_file" do
+  test "clear_key" do
     Coverband::Collectors::ViewTracker.expects(:supported_version?).returns(true)
     store = fake_store
     file_path = "#{File.expand_path(Coverband.configuration.root)}/file"
     store.raw_store.expects(:hdel).with(tracker_key, file_path)
     tracker = Coverband::Collectors::ViewTracker.new(store: store, roots: "dir")
-    tracker.track_views("name", "start", "finish", "id", identifier: file_path)
-    tracker.clear_file!("file")
-    assert_equal [], tracker.logged_views
+    tracker.track_key(identifier: file_path)
+    tracker.clear_key!("file")
+    assert_equal [], tracker.logged_keys
   end
 
   protected
