+++ views/view_tracker.erb
@@ -22,18 +22,18 @@
             <a href='<%= "#{base_path}view_tracker_data" %>'>view tracker data</a>
           <% end %>
         </h4>
-        <h2>Unused Views: (<%= tracker.unused_views.length %>)</h2>
+        <h2>Unused Views: (<%= tracker.unused_keys.length %>)</h2>
         <p>These views have not been rendered since recording started at <%= tracker.tracking_since %></p>
         <ul>
-        <% tracker.unused_views.each do |view_file| %>
+        <% tracker.unused_keys.each do |view_file| %>
           <li class="unused-views"><%= view_file %></li>
         <% end %>
         </ul>
 
-        <h2>Used Views: (<%= tracker.used_views.length %>)</h2>
+        <h2>Used Views: (<%= tracker.used_keys.length %>)</h2>
         <p>These views have been rendered at least once</p>
         <ul>
-        <% tracker.used_views.each_pair do |view_file, time_at| %>
+        <% tracker.used_keys.each_pair do |view_file, time_at| %>
           <li class="used-views">
             <%= view_file %>
             <span class="last_seen_at">last activity recorded <%= Time.at(time_at.to_i)%></span>
